add notification feature with WebSocket support; implement notificati…

+++ ez_tro_api/src/main/java/carevn/luv2code/ez_tro/dto/response/NotificationResponse.java
@@ -10,7 +10,7 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class NotificationResponse {
-    Long id;
+    Integer id;
 
     String title;
